rename files to match with main.
